Move generic method to common.py
